Update submit_jobs.py
added the "fudge factor" to include "before and after" raw data so that the trigger overlap files are complete. the job submission script is modified to include those other "before and after" raw data files in the submission. run_container is modified to remove those files after they are produced within the singularity environment. grid_job doesn't need to be modified since it is moving all* in the environment, then all* out.