Update submit_jobs.py
Removed the need to produce a beamdb file on the gpvm's and place it in the input directory prior to submitting the job. instead the ANNIE SQL text file is required in the input directory and sent in the job submission. Within the job, the BeamFetcherConfig file will be created and added to the BeamFetcher toolchain with the necessary run information. The SQL text file will be placed in the ToolAnalysis directory.