Update auto_submit_job.py
When submitting an entire run at once, it was noticed that the "fudge factor" wasn't being set correctly since the if statement was based on if the first and last part file specified were 0 and the final part file. As a result, every job submission (if done in chunks and the user to selected submit the entire run) did not grab the before/after part files for the trigger overlap. This has been changed so that it should now grab the correct raw data files for submission regardless if you process the full run or not.