Update submit_jobs.py
Alter the way we treat the overlap files. Before I was simply building the processed files with the same my_files.txt as the PreProcess TC. I was therefore building (depending on the submission files) as many as 2 extra part files because i was using the my_files.txt that included the overlap files. The previous version remedies this by just deleting the final part files so that the output part files allgn with the part files you are expecting. The new change includes creating a new my_files.txt separate for the Event Building TC that will not contain those overlap files.