Update auto_submit_job.py
Fixed the overlap issue. Before, it was just doing it for jobs that processed the entire run. Now it should work if you are just sending a specific number of part files. Also added user input whether to include overlap or not (helpful for source runs, for example where you may not want overlap files)